Upgrade CameraX and TensorFlow Lite dependencies
This commit updates the versions of CameraX and TensorFlow Lite
dependencies in the `gradle/libs.versions.toml` file.

CameraX dependencies (`cameraLifecycle`, `cameraCamera2`, `cameraView`)
are upgraded from version `1.3.4` to `1.5.0`.

TensorFlow Lite dependencies (`tensorflowLite`, `tensorflowLiteSupport`,
`tensorflowLiteGpu`) are upgraded to versions `2.17.0`, `0.5.0`, and
`2.17.0` respectively.

Additionally, a comment is added to clarify the purpose of the
`kotlin-compose` plugin.